Seed user_id in Listings table

+++ routes/web.php
@@ -22,34 +22,36 @@ Route::get('/', [ListingController::class, 'index']);
 
 
 // Show Create Form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 // Store Listing Data
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 // Single listing - Eloquent Models (check the route)
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 // Show edit Form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // Update Listing
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // Delete Listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 // Show Register Form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 // Create New User
 Route::post('/users', [UserController::class, 'store']);
 
 // Log User Out
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 
 // Show Log In form
-Route::get('/login', [UserController:: class, 'login']);
+Route::get('/login', [UserController:: class, 'login'])->name('login')->middleware('guest');
+// '->name('login')' indicates to the 'middleware('auth')' the 'name' to which redirect
+// in /Middleware/Authenticate
 
 // Log in User
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
